up/down-scale impl. x+y

+++ pkg/types/coordinate.go
@@ -2,7 +2,7 @@ package types
 
 // Coordinate is a struct for a point in a three dimensional coordinate system
 type Coordinate struct {
-	X int64 `json:"x"`
-	Y int64 `json:"y"`
-	Z int64 `json:"z"`
+	X float64 `json:"x"`
+	Y float64 `json:"y"`
+	Z float64 `json:"z"`
 }
